media-source/SourceBuffer-appendStream_event.html

<!doctype html>
<html>
<head>
  <meta charset='utf-8'>
  <title>Check event sequence of SourceBuffer#appendStream(stream)</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div>
    <video id='v' controls autoplay width='320' height='240'></video>
</div>
<div id="log"></div>

<script>
var FILE = '/media/white.webm';
var video = document.querySelector('video');
var mime = 'video/webm; codecs="vp8"';

function GET(url, processBody) {
    var xhr = new XMLHttpRequest();
    xhr.open('GET', url, true);
    xhr.responseType = 'stream';
    xhr.send();
    xhr.onload = function(e) {
        if (xhr.status != 200) {
            alert("Unexpected status code " + xhr.status + " for " + url);
            return false;
        }
        processBody(xhr.response);
    };
}
//check the browser supports the MIME used in this test
function canPlayType(mime) {
    if(video.canPlayType(mime) != 'probably') {
        this.step(function() {
            assert_unreached("Browser doesn't support the MIME used in this test.");
        });
        this.done();
        return false;
    }
    return true;
}
async_test(function(t) {
    if(!canPlayType.bind(t)(mime)) {
        return;
    }
    GET(FILE, function(stream) {
        var mediaSource = new MediaSource();
        mediaSource.addEventListener('sourceopen', function(){
            this.sourceBuffer = mediaSource.addSourceBuffer(mime);
            var events = [];
            this.sourceBuffer.addEventListener('updatestart', function() {
                events.push('updatestart');
            });
            this.sourceBuffer.addEventListener('update', function() {
                events.push('update');
            });
            this.sourceBuffer.addEventListener('updateend', function() {
                events.push('updateend');
                t.step(function() {
                    assert_array_equals(events, ['updatestart', 'update', 'updateend']);
                });
                t.done();
            });
            t.step(function() {
                assert_equals(this.sourceBuffer.updating, false);
                this.sourceBuffer.appendStream(stream);
                assert_equals(this.sourceBuffer.updating, true, 'Should be changed to true');
            }.bind(this));
        });
        video.src = window.URL.createObjectURL(mediaSource);
    });
},
'SourceBuffer#appendStream(stream) : Check the event sequence in case appending stream is succeeded.');

</script>
</body>
</html>
